requirements-mac.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
class-registry==2.1.2
configparser==3.5.0
deprecation==2.0
Django==1.11.11
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.2.0
django-dynamic-preferences==1.5.1
django-redis==4.9.0
djangorestframework==3.8.2
future==0.16.0
get==0.0.39
idna==2.6
kombu==4.1.0
mysqlclient==1.3.12
nose==1.3.7
packaging==17.1
persisting-theory==0.2.1
post==0.0.26
public==0.0.65
PyMySQL==0.8.0
pyparsing==2.2.0
python-telegram-bot==10.0.2
pytz==2018.3
query-string==0.0.28
redis==2.10.6
requests==2.18.4
six==1.11.0
telegram==0.0.1
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
